+++ packages/popover/src/Component.tsx
@@ -13,18 +13,21 @@ import { usePopper } from 'react-popper';
 import { CSSTransition } from 'react-transition-group';
 import { CSSTransitionProps } from 'react-transition-group/CSSTransition';
 import { ResizeObserver as ResizeObserverPolyfill } from '@juggle/resize-observer';
-import { BasePlacement, ModifierArguments, Obj, VariationPlacement } from '@popperjs/core';
+import { BasePlacement, Modifier, Obj, VariationPlacement } from '@popperjs/core';
 import cn from 'classnames';
 import maxSize from 'popper-max-size-modifier';
 
 import { Portal } from '@alfalab/core-components-portal';
+import { isFn } from '@alfalab/core-components-shared';
 import { Stack } from '@alfalab/core-components-stack';
+import { stackingOrder } from '@alfalab/core-components-stack-context';
 import { useLayoutEffect_SAFE_FOR_SSR } from '@alfalab/hooks';
-import { stackingOrder } from '@alfalab/stack-context';
 
 import styles from './index.module.css';
 
 type RefElement = HTMLElement | null;
+
+type ModifierOptions<T> = T extends Modifier<unknown, infer V> ? V : never;
 
 export type Position = BasePlacement | VariationPlacement;
 
@@ -66,9 +69,9 @@ export type PopoverProps = {
     preventOverflow?: boolean;
 
     /**
-     *  Позволяет поповеру подствраивать свою высоту под границы экрана, если из-за величины контента он выходит за рамки видимой области экрана
+     * Позволяет поповеру подствраивать свою высоту под границы экрана/элемента, если из-за величины контента он выходит за рамки области экрана/элемента
      */
-    availableHeight?: boolean;
+    availableHeight?: boolean | (() => Element | null | undefined);
 
     /**
      * Если `true`, будет отрисована стрелочка
@@ -143,6 +146,18 @@ export type PopoverProps = {
      * Контент
      */
     children?: ReactNode;
+
+    /**
+     * Класс для скролл контейнера
+     */
+    scrollableContentClassName?: string;
+
+    /**
+     * Минимальное расстояние стрелки до края поповера
+     *
+     * @default 24
+     */
+    arrowToEdgeMinDistance?: number;
 };
 
 const DEFAULT_TRANSITION: PopoverProps['transition'] = {
@@ -156,56 +171,34 @@ const CSS_TRANSITION_CLASS_NAMES = {
     exitActive: styles.exitActive,
 };
 
-const availableHieghtModifier = {
-    name: 'availableHeight',
-    enabled: true,
-    phase: 'beforeWrite',
-    requires: ['maxSize'],
-    fn({
-        state: {
-            modifiersData,
-            elements: { popper },
-        },
-    }: ModifierArguments<Obj>) {
-        const { height } = modifiersData.maxSize;
-
-        const content = popper.querySelector(`.${styles.scrollableContent}`) as HTMLElement;
-
-        if (content && !content.style.maxHeight) {
-            content.style.maxHeight = `${height}px`;
-        }
-    },
-};
-
 const DEFAULT_OFFSET = [0, 0];
 
-// Минимальное расстояние стрелки до края поповера
-const MIN_DISTANCE_TO_EDGE = 24;
+function createGetArrowPadding(minDistanceToEdge: number) {
+    return function getArrowPadding({
+        placement,
+    }: {
+        popper: { height: number; width: number };
+        reference: { height: number; width: number };
+        placement: Position;
+    }) {
+        if (placement === 'right-end' || placement === 'left-end') {
+            return { top: minDistanceToEdge, right: 0, bottom: 0, left: 0 };
+        }
 
-function getArrowPadding({
-    placement,
-}: {
-    popper: { height: number; width: number };
-    reference: { height: number; width: number };
-    placement: Position;
-}) {
-    if (placement === 'right-end' || placement === 'left-end') {
-        return { top: MIN_DISTANCE_TO_EDGE, right: 0, bottom: 0, left: 0 };
-    }
+        if (placement === 'top-start' || placement === 'bottom-start') {
+            return { top: 0, right: minDistanceToEdge, bottom: 0, left: 0 };
+        }
 
-    if (placement === 'top-start' || placement === 'bottom-start') {
-        return { top: 0, right: MIN_DISTANCE_TO_EDGE, bottom: 0, left: 0 };
-    }
+        if (placement === 'right-start' || placement === 'left-start') {
+            return { top: 0, right: 0, bottom: minDistanceToEdge, left: 0 };
+        }
 
-    if (placement === 'right-start' || placement === 'left-start') {
-        return { top: 0, right: 0, bottom: MIN_DISTANCE_TO_EDGE, left: 0 };
-    }
+        if (placement === 'top-end' || placement === 'bottom-end') {
+            return { top: 0, right: 0, bottom: 0, left: minDistanceToEdge };
+        }
 
-    if (placement === 'top-end' || placement === 'bottom-end') {
-        return { top: 0, right: 0, bottom: 0, left: MIN_DISTANCE_TO_EDGE };
-    }
-
-    return 0;
+        return 0;
+    };
 }
 
 export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
@@ -232,12 +225,17 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
             fallbackPlacements,
             preventOverflow = true,
             availableHeight = false,
+            scrollableContentClassName,
+            arrowToEdgeMinDistance = 24,
         },
         ref,
     ) => {
         const [referenceElement, setReferenceElement] = useState<RefElement>(anchorElement);
         const [popperElement, setPopperElement] = useState<RefElement>(null);
         const [arrowElement, setArrowElement] = useState<RefElement>(null);
+        const [maxSizeOptions, setMaxSizeOptions] = useState<
+            Partial<ModifierOptions<typeof maxSize>>
+        >(() => ({}));
 
         const updatePopperRef = useRef<() => void>();
 
@@ -252,7 +250,7 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
                     name: 'arrow',
                     options: {
                         element: arrowElement,
-                        padding: getArrowPadding,
+                        padding: createGetArrowPadding(arrowToEdgeMinDistance),
                     },
                 });
             }
@@ -269,9 +267,8 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
                 modifiers.push({ name: 'preventOverflow', options: { mainAxis: false } });
             }
 
-            if (availableHeight) {
-                modifiers.push({ ...maxSize, options: {} });
-                modifiers.push({ ...availableHieghtModifier, options: {} });
+            if (isFn(availableHeight) || availableHeight) {
+                modifiers.push({ ...maxSize, options: maxSizeOptions });
             }
 
             return modifiers;
@@ -283,12 +280,15 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
             preventOverflow,
             availableHeight,
             arrowElement,
+            arrowToEdgeMinDistance,
+            maxSizeOptions,
         ]);
 
         const {
             styles: popperStyles,
             attributes,
             update: updatePopper,
+            state,
         } = usePopper(referenceElement, popperElement, {
             placement: position,
             modifiers: popperModifiers,
@@ -297,6 +297,14 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
         if (updatePopper) {
             updatePopperRef.current = updatePopper;
         }
+
+        useLayoutEffect_SAFE_FOR_SSR(() => {
+            const nextBoundry = isFn(availableHeight) ? availableHeight() ?? undefined : undefined;
+
+            if (!(maxSizeOptions.boundary === nextBoundry)) {
+                setMaxSizeOptions({ boundary: nextBoundry });
+            }
+        });
 
         useLayoutEffect_SAFE_FOR_SSR(() => {
             setReferenceElement(anchorElement);
@@ -347,7 +355,17 @@ export const Popover = forwardRef<HTMLDivElement, PopoverProps>(
                 {...attributes.popper}
             >
                 <div className={cn(styles.inner, popperClassName)} ref={innerRef}>
-                    <div className={cn({ [styles.scrollableContent]: availableHeight })}>
+                    <div
+                        className={cn(scrollableContentClassName, {
+                            [styles.scrollableContent]: isFn(availableHeight) || availableHeight,
+                        })}
+                        style={{
+                            maxHeight:
+                                isFn(availableHeight) || availableHeight
+                                    ? state?.modifiersData.maxSize?.height
+                                    : undefined,
+                        }}
+                    >
                         {children}
                     </div>