+++ packages/universal-modal/src/components/content/responsive/Component.responsive.tsx
@@ -2,10 +2,10 @@ import React, { FC } from 'react';
 
 import { useIsDesktop } from '@alfalab/core-components-mq';
 
-import { ContentDesktop, ContentDesktopProps } from '../desktop/Component.desktop';
-import { ContentMobile, ContentMobileProps } from '../mobile/Component.mobile';
+import { type ContentDesktopProps, ContentDesktop } from '../desktop/Component.desktop';
+import { type ContentMobileProps, ContentMobile } from '../mobile/Component.mobile';
 
-export type ContentResponsiveProps = ContentDesktopProps | ContentMobileProps;
+export interface ContentResponsiveProps extends ContentDesktopProps, ContentMobileProps {}
 
 export const ContentResponsive: FC<ContentResponsiveProps> = ({ ...restProps }) => {
     const isDesktop = useIsDesktop();