+++ typings/environment.d.ts
@@ -1,9 +1,0 @@
-declare global {
-    namespace NodeJS {
-        interface ProcessEnv {
-            NODE_ENV: 'development' | 'production' | 'test';
-        }
-    }
-}
-
-export {};