+++ typings/libphonenumber-js.d.ts
@@ -1,5 +1,0 @@
-declare module 'libphonenumber-js/bundle/libphonenumber-js.min' {
-    import { AsYouType } from 'libphonenumber-js';
-
-    export { AsYouType };
-}