+++ packages/date-input/CHANGELOG.md
@@ -1,4 +1,58 @@
-# Change Log
+# @alfalab/core-components-date-input
+
+## 5.0.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-input@16.0.1
+
+## 5.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-input@16.0.0
+
+## 4.4.20
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.6.2
+
+## 4.4.19
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.6.1
+
+## 4.4.18
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.6.0
+
+## 4.4.17
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   input@15.5.5
 
 ## 4.4.16