+++ packages/markdown/CHANGELOG.md
@@ -1,5 +1,95 @@
 # @alfalab/core-components-markdown
 
+## 2.0.1
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-mq@5.0.1
+    -   @alfalab/core-components-typography@5.0.1
+    -   @alfalab/core-components-list@4.0.1
+
+## 2.0.0
+
+### Major Changes
+
+<sup><time>05.08.2025</time></sup>
+
+#### [#1611](https://github.com/core-ds/core-components/pull/1611)
+
+Обновлена сборка.
+
+Добавлены пропущенные зависимости.
+
+Синхронизированы версии зависимостей.
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   @alfalab/core-components-link@6.0.0
+    -   @alfalab/core-components-list@4.0.0
+    -   @alfalab/core-components-mq@5.0.0
+    -   @alfalab/core-components-typography@5.0.0
+
+## 1.7.0
+
+### Minor Changes
+
+<sup><time>18.07.2025</time></sup>
+
+### [#1765](https://github.com/core-ds/core-components/pull/1765)
+
+-   Заменено длинное тире на среднее "–" в маркерах списков
+-   Изменено выравнивание текста в контейнере с маркером на правое с сохранением ширины
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   list@3.6.0
+
+## 1.6.16
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.18.0
+    -   typography@4.14.3
+    -   list@3.5.7
+
+## 1.6.15
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.1
+    -   typography@4.14.2
+    -   list@3.5.6
+
+## 1.6.14
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.14.1
+    -   list@3.5.5
+
+## 1.6.13
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   typography@4.14.0
+    -   list@3.5.4
+
+## 1.6.12
+
+### Patch Changes
+
+-   Обновлены зависимости
+    -   shared@0.17.0
+    -   typography@4.13.1
+    -   list@3.5.3
+
 ## 1.6.11
 
 ### Patch Changes